Fix linker conflicts for GUI globals

+++ iscalerh2pf/ImageProc/iChips/C341/driver/dvC341_OSD.c
@@ -1,4 +1,5 @@
 //#include <string.h>
+#include <stdint.h>
 
 #include "dvC341.h"
 #include "dvC341_OSD.h"
@@ -973,7 +974,7 @@ void dvC341_SerialFlash_Write(UINT32 ulSflAd, UINT8 *pucBuffer, UINT32 ulDataSiz
     UINT8*  pucData = pucBuffer;
     UINT32  ulTotoalCheckSum = 0;
     UINT32  ulCheckSum = 0;
-    UINT32  ulReadFlashAdd = (UINT32)pucBuffer;
+    uintptr_t ulReadFlashAdd = (uintptr_t)pucBuffer;
     UINT32  ulSectorSize = 0; //A70LV_Larry_0158
 
     //matching 64k
